+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Changed for changes in existing functionality.
 Deprecated for soon-to-be removed features.
 Removed for now removed features.
 Fixed for any bug fixes.
-Security in case of vulnerabilities. 
+Security in case of vulnerabilities.
 -->
 
 ## [0.2.0] - 2025-01-01
@@ -20,10 +20,11 @@ Security in case of vulnerabilities.
 
 - `Changed`: Add sourcemaps and source typescript to the npm pack
 
+- `Fixed`: macOS system mount points are now filtered properly
+
 ## [0.1.0] - 2024-12-17
 
 First release! Everything is a new feature!
 
 The 1.0.0 release will happen after some integration testing with the native
 library payloads, but the API should be stable after the first release.
-